requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.30.0
streamlit>=1.25.0
pillow>=9.0.0
accelerate>=0.20.0 